+++ cider-util.el
@@ -135,8 +135,9 @@ find a symbol if there isn't one at point."
         (unless (text-property-any 0 (length str) 'field 'cider-repl-prompt str)
           ;; Remove font-locking, prefix quotes, and trailing . from constructors like Record.
           (thread-last (substring-no-properties str)
-            ;; constructurs (Foo.)
+            ;; constructors (Foo.)
             (string-remove-suffix ".")
+            ;; quoted symbols ('sym)
             (string-remove-prefix "'")
             ;; var references (#'inc 2)
             (string-remove-prefix "#'"))))