+++ cider-util.el
@@ -313,6 +313,9 @@ This buffer is not designed to display anything to the user.  For that, use
         (with-current-buffer (cider--make-buffer-for-mode mode)
           (erase-buffer)
           (insert string)
+          ;; don't try to font-lock unbalanced Clojure code
+          (when (eq mode 'clojure-mode)
+            (check-parens))
           (font-lock-fontify-region (point-min) (point-max))
           (buffer-string))
       string)))
@@ -329,7 +332,11 @@ Unless you specify a BUFFER it will default to the current one."
 
 (defun cider-font-lock-as-clojure (string)
   "Font-lock STRING as Clojure code."
-  (cider-font-lock-as 'clojure-mode string))
+  ;; If something goes wrong (e.g. the code is not balanced)
+  ;; we simply return the string.
+  (condition-case nil
+      (cider-font-lock-as 'clojure-mode string)
+    (error string)))
 
 ;; Button allowing use of `font-lock-face', ignoring any inherited `face'
 (define-button-type 'cider-plain-button