+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [#2828](https://github.com/clojure-emacs/cider/pull/2828): Bind "L" to toggle display of locals during a debug session.
 * [#2800](https://github.com/clojure-emacs/cider/pull/2800): Add support for force-out debugger command.
 * Add support for nREPL 0.8 completion op. It's used if `cider-nrepl` is not available.
+* Add the browser type to ClojureScript REPL types.
 
 ### Changes